[CLEANUP] Moved fallback fonts initializstion to fontman class + random logging cleanups
